move: common > ApplyPage 하위로 이동
