+++ src/views/ApplyPage/components/FileInput/icons/IconPlusButton.tsx
